Move GatherFusionPattern into elementwise fusion
