+++ compiler/src/iree/compiler/DispatchCreation/test/BUILD.bazel
@@ -21,7 +21,7 @@ iree_lit_test_suite(
             "collapse_linalg_generic_on_tensors.mlir",
             "collapse_reduction.mlir",
             "attention_fuse_by_expansion.mlir",
-            "fold_transpose.mlir",
+            "elementwise_op_fusion.mlir",
             "dispatch_linalg_transform_dialect.mlir",
             "dispatch_region_formation_preprocessing.mlir",
             "fold_unit_dims.mlir",
