+++ compiler/src/iree/compiler/DispatchCreation/test/CMakeLists.txt
@@ -27,7 +27,7 @@ iree_lit_test_suite(
     "dispatch_linalg_on_tensors_fusion_with_transpose.mlir"
     "dispatch_linalg_transform_dialect.mlir"
     "dispatch_region_formation_preprocessing.mlir"
-    "fold_transpose.mlir"
+    "elementwise_op_fusion.mlir"
     "fold_unit_dims.mlir"
     "form_dispatch_regions.mlir"
     "form_dispatch_workgroups.mlir"
